move css to group dir
